+++ src/main/java/dev/robert/spring_boot/animal_shelter_spring/repository/AdoptionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.robert.spring_boot.animal_shelter_spring.model.Adoption;
 
-public interface AdoptionRepository extends JpaRepository<Adoption, Integer>{
+public interface AdoptionRepository extends JpaRepository<Adoption, Long>{
 
 }
