+++ src/main/java/dev/robert/spring_boot/animal_shelter_spring/repository/AnimalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.robert.spring_boot.animal_shelter_spring.model.Animal;
 
-public interface AnimalRepository extends JpaRepository<Animal, Integer>{
+public interface AnimalRepository extends JpaRepository<Animal, Long>{
 
 }
